Add dependency matrix in main.yml to test modules in their independent environment (#738)
* test forte models

* test forte models

* test extra format

* extra models dependency

* format

* rm src/

* format

* format

* format

* format

* Format

* Format

* test models

* install models deps

* matrix modules

* matrix modules

* model modules

* test new module format

* get module name

* get module name

* get module name

* get module name

* module name

* module name

* module name

* models module

* models module

* models module

* models module

* models module

* models module

* dependency matrix

* move testing modules down

* test

* test

* pytest

* modules

* ir test folder

* correct extractor test files

* move test modules later

* install latestest texar

* module empty case

* remove script to remove texar pytorch

* install texar after texar nondependency test

* Update setup.py

* Update main.yml

* texar 0.1.4

* rearrange test order in main.yml

* rearrange module orders in setuppy and README.md

* nondependency test comments

* fix test module and test files in main.yml

* fix test file and tagging dependency in README.md

* add tests for remote processor

* indentation level

* move types has extra dependency out

* remove shortcut path for Pipeline and BaseExtractor

* remove commented imports

* remove classes have dependency on transformers

* adjust workflow configuration

* fix MatrixLike importing

* fix slice_batch importing

* move test cases for RemoteProcessor out

* Remove unnecessary comments

* remove pipeline import in __init__.py

* fix importing paths

* added remote_processor_test.py into main.yml and fix test in advanced_pipeline_test.py

* revert types.py and try catch for MatrixLike package importing

* revert some relative paths changes and correct the testing order

* temporarily add code to install torch

* pylint

* spelling

* correct testing order

* pylint

* Revert __init__.py

* from forte import pipeline

* revert relative path imports

* align module order in main.yml and setup.py

* revert the order of spelling word

* revert imports in test cases

* separate test main and modules

* test new condition syntax

* test new condition syntax

* test new condition syntax

* test new condition syntax

* test new condition syntax

* reert machine_translation_processor_test.py

* add new job for testing modules

* minimize env pacakge versions for testing modules

* minimize env pacakge versions for testing modules

* remove notebook details

* fix testing code in test module

* no test extra modules temporarily

* correct changes

* remove coverage installation

* docs/requirements.txt

* revert bank customer intent

* update README.md from master

* update README.md from master

* update docs/requirements.txt from master

* test extra modules with pytorch installed

* install pytest

* install tensorflow temporarily

* install tensorflow temporarily

* add tensorflow into ir module and add libsndfile1-dev installation

* add stave installation for testing modules

* relax tensorflow version requirement

* add pip and django installation for job of test_modules

* add tensorflow installation for module models
