+++ tests/forte/processors/ir/bert_based_query_creator_test.py
